Implement HybridCache
